Refactorings
Removed budget entity, now everything hangs from a month
Adjusted IDs to be all Long
Adjusted annotations for table and entity names
Added Controller for easy to read data
Added endpoint for aggregated month spendings

+++ src/main/java/com/budgetApp/crud/category/CategoryController.java
@@ -6,7 +6,7 @@ import org.springframework.web.bind.annotation.RestController;
 
 @RestController
 @RequestMapping("/category")
-public class CategoryController extends AbstractCRUDController<Category, Integer, CategoryRepository> {
+public class CategoryController extends AbstractCRUDController<Category, Long, CategoryRepository> {
 
 
     public CategoryController(CategoryRepository repository) {
