+++ src/main/java/com/budgetApp/crud/category/CategoryRepository.java
@@ -2,5 +2,5 @@ package com.budgetApp.crud.category;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface CategoryRepository extends CrudRepository<Category, Integer> {
+public interface CategoryRepository extends CrudRepository<Category, Long> {
 }
