+++ src/main/java/com/budgetApp/crud/category/SubcategoryRepository.java
@@ -2,5 +2,5 @@ package com.budgetApp.crud.category;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface SubcategoryRepository extends CrudRepository<Subcategory, Integer> {
+public interface SubcategoryRepository extends CrudRepository<Subcategory, Long> {
 }
